Update README.md
Create A-Prueba.postman_collection.json

Update README.md

Update README.md

+++ Application/Features/OrganizacionesProductos/Commands/DeleteOrganizacionCommandHandler.cs
@@ -23,10 +23,23 @@ namespace Application.Features.OrganizacionesProductos.Commands
 
         public async Task<int> Handle(DeleteOrganizacionCommand request, CancellationToken cancellationToken)
         {
-            var organizacionToDelete = await _organizacionProductoRepository.GetByID(request.IdOrganizacion);
+
+            var validator = new DeleteOrganizacionCommandValidator();
+            var validationResult = await validator.ValidateAsync(request);
+            if (validationResult.Errors.Count > 0) throw new Common.ValidationException(validationResult);
+
+
+            var organizacionToDelete = _organizacionProductoRepository.GetByID(request.IdOrganizacion);
+
+            if (organizacionToDelete == null)
+            {
+                // La entidad no existe
+                throw new Exception($"No se encontró la organización con ID {request.IdOrganizacion}");
+            }
+
             await _organizacionProductoRepository.DeleteAsync(organizacionToDelete);
 
-            return 0;
+            return 1;
         }
     }
 }
