+++ Application/Features/OrganizacionesUsuarios/Commands/DeleteOrganizacionCommandHandler.cs
@@ -24,7 +24,18 @@ namespace Application.Features.OrganizacionesUsuarios.Commands
 
         public async Task<int> Handle(DeleteOrganizacionCommand request, CancellationToken cancellationToken)
         {
-            var organizacionToDelete = await _organizacionUsuarioRepository.GetByID(request.IdOrganizacion);
+            var validator = new DeleteOrganizacionCommandValidator();
+            var validationResult = await validator.ValidateAsync(request);
+            if (validationResult.Errors.Count > 0) throw new Common.ValidationException(validationResult);
+
+            var organizacionToDelete = _organizacionUsuarioRepository.GetByID(request.IdOrganizacion);
+
+            if (organizacionToDelete == null)
+            {
+                // La entidad no existe, puedes lanzar una excepción adecuada
+                throw new Exception($"No se encontró la organización con ID {request.IdOrganizacion}");
+            }
+
             await _organizacionUsuarioRepository.DeleteAsync(organizacionToDelete);
 
             return 0;
