+++ Application/Features/Productos/Commands/DeleteProductoCommandHandler.cs
@@ -23,10 +23,22 @@ namespace Application.Features.Productos.Commands
 
         public async Task<int> Handle(DeleteProductoCommand request, CancellationToken cancellationToken)
         {
+            var validator = new DeleteProductoCommandValidator();
+            var validationResult = await validator.ValidateAsync(request);
+            if (validationResult.Errors.Count > 0) throw new Common.ValidationException(validationResult);
+
             var productoToDelete = await _productoRepository.GetByID(request.IdProducto);
+
+
+            if (productoToDelete == null)
+            {
+                
+                throw new Exception($"No se encontró el producto con ID {request.IdProducto}");
+            }
+
             await _productoRepository.DeleteAsync(productoToDelete);
 
-            return 0;
+            return 1;
         }
     }
 }
